feat(ui): improve the display of unregistered workflow nodes
