+++ ui/src/components/workflow/designer/nodes/index.ts
@@ -1,29 +1,3 @@
-﻿import { BizApplyNodeRegistry } from "./BizApplyNodeRegistry";
-import { BizDeployNodeRegistry } from "./BizDeployNodeRegistry";
-import { BizMonitorNodeRegistry } from "./BizMonitorNodeRegistry";
-import { BizNotifyNodeRegistry } from "./BizNotifyNodeRegistry";
-import { BizUploadNodeRegistry } from "./BizUploadNodeRegistry";
-import { BranchBlockNodeRegistry, ConditionNodeRegistry } from "./ConditionNode";
-import { DelayNodeRegistry } from "./DelayNode";
-import { EndNodeRegistry } from "./EndNode";
-import { StartNodeRegistry } from "./StartNode";
-import { CatchBlockNodeRegistry, TryCatchNodeRegistry } from "./TryCatchNode";
-
-export const getAllNodeRegistries = () => {
-  return [
-    StartNodeRegistry,
-    EndNodeRegistry,
-    DelayNodeRegistry,
-    BizApplyNodeRegistry,
-    BizUploadNodeRegistry,
-    BizMonitorNodeRegistry,
-    BizDeployNodeRegistry,
-    BizNotifyNodeRegistry,
-    ConditionNodeRegistry,
-    BranchBlockNodeRegistry,
-    TryCatchNodeRegistry,
-    CatchBlockNodeRegistry,
-  ];
-};
+﻿export { getAllNodeRegistries, isNodeTypeRegistered } from "./NodeRegistries";
 
 export type * from "./typings";
